requirements.txt
# Backend - FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM - HuggingFace
transformers==4.37.0
torch==2.1.2
accelerate==0.26.0
sentencepiece==0.1.99
protobuf==4.25.2

# Frontend - Streamlit
streamlit==1.31.0

# RAG - Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.2.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1

# Development
jupyter==1.0.0
ipykernel==6.29.0
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.1.1
